Switch to libdeflate
